requirements.txt
flask==2.2.2
flask-restful==0.3.9
pytest==7.2.0
pytest-mock==3.10.0
PyYAML==6.0
mypy==0.991
autopep8==2.0.0
requests==2.28.1
redis==4.3.4

types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-redis==4.3.21.4